# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "chat_backend",
    "version": "v1.0"
  },
  "paths": {
...
    "/persons": {
      "get": {
        "tags": [
          "Persons"
        ],
        "summary": "Возвращает всех военнослужащих, пользующихся чатом",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Person"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Person"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Person"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Persons"
        ],
        "summary": "Создаёт запись о военнослужащем, пользующемся чатом\r\nДоступно только администраторам",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePerson"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePerson"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePerson"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ValidationFailure"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ValidationFailure"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ValidationFailure"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/persons/by-guid": {
      "get": {
        "tags": [
          "Persons"
        ],
        "summary": "Возвращает военнослужащего, пользующегося чатом, с указанным Guid",
        "parameters": [
          {
            "name": "guid",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Person"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Person"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Person"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateFullName": {
        "required": [
          "firstName",
          "lastName",
          "parentName"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "maxLength": 500,
            "minLength": 1,
            "type": "string",
            "description": "Имя",
            "example": "Иван"
          },
          "lastName": {
            "maxLength": 500,
            "minLength": 1,
            "type": "string",
            "description": "Фамилия",
            "example": "Иванов"
          },
          "parentName": {
            "maxLength": 500,
            "minLength": 1,
            "type": "string",
            "description": "Отчество",
            "example": "Иванович"
          }
        },
        "additionalProperties": false,
        "description": "Создаёт ФИО"
      },
...
      "CreatePerson": {
        "required": [
          "fullName",
          "militaryRankId"
        ],
        "type": "object",
        "properties": {
          "fullName": {
            "$ref": "#/components/schemas/CreateFullName"
          },
          "militaryRankId": {
            "type": "integer",
            "description": "Id воинского звания",
            "format": "int32",
            "example": 1
          }
        },
        "additionalProperties": false,
        "description": "Создаёт запись о военнослужащем"
      },
...
      "FullName": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "Имя",
            "nullable": true,
            "readOnly": true
          },
          "lastName": {
            "type": "string",
            "description": "Фамилия",
            "nullable": true,
            "readOnly": true
          },
          "parentName": {
            "type": "string",
            "description": "Отчество",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "description": "ФИО"
      },
...
      "MilitaryRank": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Id воинского звания",
            "format": "int32",
            "readOnly": true
          },
          "fullName": {
            "type": "string",
            "description": "Полное название звания",
            "nullable": true,
            "readOnly": true
          },
          "shortName": {
            "type": "string",
            "description": "Сокращенное название звания",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "description": "Воинское звание"
      },
...
      "Person": {
        "type": "object",
        "properties": {
          "guid": {
            "type": "string",
            "description": "Guid военнослужащего",
            "format": "uuid",
            "readOnly": true
          },
          "fullName": {
            "$ref": "#/components/schemas/FullName"
          },
          "rank": {
            "$ref": "#/components/schemas/MilitaryRank"
          }
        },
        "additionalProperties": false,
        "description": "Военнослужащий"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "Severity": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "ValidationFailure": {
        "type": "object",
        "properties": {
          "propertyName": {
            "type": "string",
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "attemptedValue": {
            "nullable": true
          },
          "customState": {
            "nullable": true
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "errorCode": {
            "type": "string",
            "nullable": true
          },
          "formattedMessagePlaceholderValues": {
            "type": "object",
            "additionalProperties": { },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "jwt_auth": {
        "type": "http",
        "description": "Please insert JWT with Bearer into field",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "jwt_auth": [ ]
    }
  ]
}